pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "tex"
version = "0.1.0"
description = "A transparent, local, explainable Linux task assistant"
readme = "README.md"
requires-python = ">=3.10"
license = { text = "MIT" }
dependencies = [
    "typer>=0.12",
    "rich>=13.0",
    "pydantic>=2.0",
    "ollama>=0.2",
    "httpx>=0.27",
    "python-dotenv>=1.0",
    "loguru>=0.7",
]

[project.scripts]
tex = "tex.main:app"

[project.optional-dependencies]
dev = [
    "ruff",
    "pytest",
]

[tool.ruff]
line-length = 100
target-version = "py310"
